--- a/examples/yolov8.py
+++ b/examples/yolov8.py
@@ -331,25 +331,25 @@ def compute_iou_matrix(boxes):
   iou = intersection / (areas[:, None] + areas[None, :] - intersection)
   return iou
 
-def postprocess(output,max_det=300):
-  xc, yc, w, h, class_scores = output[0][0], output[0][1], output[0][2], output[0][3], output[0][4:]
-  class_ids = Tensor.argmax(class_scores,axis=0)
-  probs = Tensor.max(class_scores, axis=0)
-  probs = Tensor.where(probs>=0.25,probs,0)
-  x1 = xc - w / 2
-  y1 = yc - h / 2
-  x2 = xc + w / 2
-  y2 = yc + h / 2
-  boxes = Tensor.stack(x1, y1, x2, y2, probs, class_ids, dim=1)
-  order = Tensor.topk(probs,max_det)[1]
-  boxes = boxes[order]
-  iou = compute_iou_matrix(boxes)
-  iou = Tensor.triu(iou,diagonal=1)
-  high_iou_mask = iou > 0.45
-  overlap_counts = high_iou_mask.sum(axis=0)
-  no_overlap_mask = overlap_counts == 0
-  boxes = boxes * no_overlap_mask.unsqueeze(-1)
-  return boxes
+def postprocess(output, max_det=300):
+    xc, yc, w, h, class_scores = output[0][0], output[0][1], output[0][2], output[0][3], output[0][4:]
+    class_ids = Tensor.argmax(class_scores, axis=0)
+    probs = Tensor.max(class_scores, axis=0)
+    probs = Tensor.where(probs >= 0.25, probs, 0)
+    x1 = xc - w / 2
+    y1 = yc - h / 2
+    x2 = xc + w / 2
+    y2 = yc + h / 2
+    boxes = Tensor.stack(x1, y1, x2, y2, probs, class_ids, dim=1)
+    order = Tensor.topk(probs, max_det)[1]
+    boxes = boxes[order]
+    iou = compute_iou_matrix(boxes[:, :4])
+    iou = Tensor.triu(iou, diagonal=1)
+    same_class_mask = boxes[:, -1][:, None] == boxes[:, -1][None, :]
+    high_iou_mask = (iou > 0.45) & same_class_mask
+    no_overlap_mask = high_iou_mask.sum(axis=0) == 0
+    boxes = boxes * no_overlap_mask.unsqueeze(-1)
+    return boxes
 
 def get_weights_location(yolo_variant: str) -> Path:
   weights_location = Path(__file__).parents[1] / "weights" / f'yolov8{yolo_variant}.safetensors'