… examples/webgpu/yolov8/index.html
… test/external/external_test_yolov8.py