converted some of the post processing from numpy to tinygrad, performance is still similar.

master:  
https://roryclear.github.io/yolo-webgpu/index_master.html
https://roryclear.github.io/yolo-webgpu/index_640_master.html
|            | 416x416 | 640x640 |
|------------|---------|---------|
| M2 Mac Mini| 20fps   |   17fps |
| iPhone 13  |  3.6fps |   3.4fps|

this branch:  
https://roryclear.github.io/yolo-webgpu/index.html
https://roryclear.github.io/yolo-webgpu/index_640.html
|            | 416x416 | 640x640 |
|------------|---------|---------|
| M2 Mac Mini| 20fps   |   20fps |
| iPhone 13  | 3.3fps   |  3.5fps|