@geohot 2025-04-24
Ahh, it's nice to see the processing stuff moved into tinygrad. I'll test building the page on my mac, if that works I'll merge.