@geohot 2025-04-24
This hangs on Loading model on my Z Fold 5