@roryclear 2025-04-25
> This hangs on Loading model on my Z Fold 5, I'll tweet once it has good error messages. Merged cause that's not a regression from master.

nice, I'll try get a pr in with a loading error